added: <constant name="name_of_constant">number</constant> ability (in includes) to allow things such as <posx>name_of_constant</posx> for easy alteration of key parameters in skins.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@10266 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/VisibleEffect.h
@@ -69,8 +69,8 @@ private:
   float m_endY;
   float m_centerX;
   float m_centerY;
-  int m_startAlpha;
-  int m_endAlpha;
+  float m_startAlpha;
+  float m_endAlpha;
 
   // timing variables
   unsigned int m_start;
